Move App to Components folder. Split Store and Theme.
