+++ src/components/App/index.js
@@ -2,14 +2,13 @@ import React from 'react';
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 
 import './App.css';
-import {
-  BottomBar,
-  LearnPage,
-  TopBar,
-  DictionaryPage,
-  TodayPage,
-  UserPage
-} from './components';
+
+import BottomBar from '../BottomBar';
+import DictionaryPage from '../DictionaryPage';
+import LearnPage from '../LearnPage';
+import TodayPage from '../TodayPage';
+import TopBar from '../TopBar';
+import UserPage from '../UserPage';
 
 const App = () => {
   return (
