--- src/components/index.js
@@ -1,6 +1,0 @@
-export { default as BottomBar } from './BottomBar';
-export { default as DictionaryPage } from './DictionaryPage';
-export { default as LearnPage } from './LearnPage';
-export { default as TopBar } from './TopBar';
-export { default as UserPage } from './UserPage';
-export { default as TodayPage } from './TodayPage';
